improve database
Each feed now keeps track of a guild, so that feeds per guild can be supported, and it allows one feed per channel. The entries table was also improved to have one channel/feed pair per entry. The function for getting information for a feed were added to and modified and moved to another class.